Moved models tests out of elastiknn-testing
