+++ elastiknn-lucene/src/main/java/com/klibisz/elastiknn/lucene/HashFieldType.java
@@ -1,0 +1,16 @@
+package com.klibisz.elastiknn.lucene;
+
+import org.apache.lucene.document.FieldType;
+import org.apache.lucene.index.IndexOptions;
+
+public final class HashFieldType extends FieldType {
+
+    public HashFieldType() {
+        setTokenized(false);
+        setIndexOptions(IndexOptions.DOCS_AND_FREQS);
+        setOmitNorms(true);
+        freeze();
+    }
+
+    public static HashFieldType HASH_FIELD_TYPE = new HashFieldType();
+}
