+++ elastiknn-models/src/test/scala/com/klibisz/elastiknn/models/HammingLshModelSuite.scala
