+++ elastiknn-models/src/test/scala/com/klibisz/elastiknn/models/L2LshSuite.scala
